Fix s390_ccm() to receive mac_length and nonce_length as unsigned int, not long
The caller of s390_ccm() passes an unsigned int mac length and nonce_length
to s390_ccm(). Furthermore s390_ccm() passes them to s390_ccm_authenticate()
which uses an unsigned int.

This mismatch does not produce a problem, because the value of mac_length
and nonce_length can anyway not be larger than unsigned int, because the
outer most caller has it as unsigned int.

+++ src/include/s390_ccm.h
@@ -286,8 +286,8 @@ static inline unsigned int s390_ccm(unsigned int function_code,
 		      unsigned char *payload, unsigned long payload_length,
 		      unsigned char *ciphertext,
 		      const unsigned char *assoc_data, unsigned long assoc_data_length,
-		      const unsigned char *nonce, unsigned long nonce_length,
-		      unsigned char *mac, unsigned long mac_length,
+		      const unsigned char *nonce, unsigned int nonce_length,
+		      unsigned char *mac, unsigned int mac_length,
 		      unsigned char *key)
 {
 	unsigned char initial_ctr[AES_BLOCK_SIZE];
